+++ learn-you-a-haskell/io/posixtime/src/Main.hs
@@ -3,21 +3,27 @@ module Main where
 import System.Posix.Files
 import System.Posix.Types
 import System.Environment
+import Data.Time.Clock.POSIX
+import Data.Time.Clock
 
-getAccessTime :: FilePath -> IO (EpochTime)
+getAccessTime :: FilePath -> IO (POSIXTime)
 getAccessTime filePath =
   do stat <- getFileStatus filePath
-     return (accessTime stat)
+     return (accessTimeHiRes stat)
 
-getModificationTime :: FilePath -> IO (EpochTime)
+getModificationTime :: FilePath -> IO (POSIXTime)
 getModificationTime filePath =
   do stat <- getFileStatus filePath
-     return (modificationTime stat)
+     return (modificationTimeHiRes stat)
 
-getChangeTime :: FilePath -> IO (EpochTime)
+getChangeTime :: FilePath -> IO (POSIXTime)
 getChangeTime filePath =
   do stat <- getFileStatus filePath
-     return (statusChangeTime stat)
+     return (statusChangeTimeHiRes stat)
+
+-- Wrapper function
+posixToUTC :: POSIXTime -> UTCTime
+posixToUTC inputPosixTime = posixSecondsToUTCTime inputPosixTime
 
 main :: IO ()
 main = do
@@ -25,4 +31,4 @@ main = do
   accessTime <- getAccessTime input
   modifTime <- getModificationTime input
   changeTime <- getChangeTime input
-  print [accessTime, modifTime, changeTime]
+  putStrLn (show input ++ "\t Access: " ++ show (posixToUTC accessTime) ++ "\t Modify: " ++ show (posixToUTC modifTime) ++ "\t Change: " ++ show (posixToUTC changeTime))