+++ learn-you-a-haskell/io/posixtime/ChangeLog.md
@@ -1,5 +1,14 @@
 # Revision history for posixtime
 
+## 0.1.2 --2020-02-13
+
+* Adding a more readable print output
+
+## 0.1.1 --2020-02-12
+
+* Sub calculus for "files" time. The functions return in POSIXTime (Haskell data type)
+  and the final output is in UTCTime (Haskell data type)
+
 ## 0.1.0  -- 2020-02-11
 
 * First version. Released on an unsuspecting world.